fix - jar shading causes duplicate class warnings in maven
problem is identical to http://stackoverflow.com/questions/14402745/duplicate-classes-in-commons-collections-and-commons-beanutils. To avoid duplicate class problem, the commons-beanutils dependency was excluded from the poms of the modules that do the jar shading, as suggested here http://stackoverflow.com/questions/9802736/maven-build-warning-we-have-a-duplicate-class.

also removed not needed dependencies (log4j and aspectj related) that caused the same duplicate class warnings during shading