fix for the rpslattribute comment issue: - removed separate constructor that directly specified comment; - renamed getComment() to getCleanComment() - added WhoisObjectMapper code to merge comment from WhoisObject's Attribute into RpslAttribute's value.
Reason: in the SQL DB, we store object blobs, possibly with multiline attributes and multiple EOL comments at each line. It is a hard requirement that we have to return this in the same format as it was submitted. This means that no matter what, we always have to keep the 'dirty' value in RpslAttribute. Having an 'override comment' on top of this design is not feasible.