Updated Spring by applying AOP workaround suggested in https://github.com/spring-projects/spring-framework/issues/32882#issuecomment-2127283683
"These problems come from mixed AspectJ weaving and proxying. In your scenario, you seem to be using @EnableTransactionManagement in AspectJ mode but @EnableCaching in proxy mode. For the time being, you could try to switch @EnableCaching and any other aspect-related functionality to AspectJ mode as well and remove any variant of@EnableAspectJAutoProxy completely, if that is feasible for your application."