Added conditional attribute requirement to Object Template (#1881)
* Added conditional requirement

* On inet(6)num objects for allocations or End User assignments made by the RIPE NCC, the "org:" attribute is mandatory.

* "mnt-lower" on inet(6)num is conditional.

On an inetnum object with the status "ALLOCATED PA" or "ALLOCATED-ASSIGNED PA", the RIPE NCC has specified a "mnt-lower:" attribute. It cannot be removed and another "mnt-lower:" attribute cannot be added. The same applies to an inet6num object with the status "ALLOCATED-BY-RIR".

* "sponsoring-org" is conditional.

Only the RIPE NCC can add, change or remove the "sponsoring-org:" attribute. It must reference an organisation of type "LIR".

* organisation abuse-c is conditional.

The "abuse-c:" attribute must reference a role object that has an "abuse-mailbox:" attribute. Once set, the attribute cannot be removed if the organisation is referenced by any object that has a RIPE NCC maintainer present.

* phone: and fax-no: are conditional on organisation objects.

If the organisation object has a RIPE NCC maintainer present, only the RIPE NCC can change the value of the "org-name:" attribute after an LIR submits a change request. In addition, LIRs cannot change the "e-mail:", "phone:", and "fax-no:" value; these must be changed in the LIR Portal.

* role abuse-mailbox is conditional.

The "abuse-mailbox:" attribute cannot be removed from a role object if it's referenced as an "abuse-c:" attribute in any organisation object.

* peering-set peering: and mp-peering attributes are conditional.

A peering-set object must include either a "peering:" or an "mp-peering:" attribute. One of these must exist, but both attributes cannot be present at the same time.

* Filter-set filter: and mp-filter attributes are conditional.

A filter-set object must include either a "filter:" or an "mp-filter:" attribute. One of these must exist, but both attributes cannot be present at the same time.

* aut-num org: attribute is conditional.

If the object has a RIPE NCC maintainer as one of the maintainers, only the RIPE NCC can change the value of the "org:" attribute.

* Fixed tests